double add

+++ Models/Debits.cs
@@ -11,10 +11,10 @@ namespace PersonalFinance.Models
         public string CredName { get; set; }
         public DateTime DebDateTime { get; set; } //Scadenza debito
         public DateTime DebInsDate { get; set; }
-        public float RtNum { get; set; } //Numero di rate
-        public float RtPaid { get; set; } //Rate pagate
-        public float RemainToPay { get; set; } //Importo da pagare
-        public float DebValue { get; set; }
+        public double RtNum { get; set; } //Numero di rate
+        public double RtPaid { get; set; } //Rate pagate
+        public double RemainToPay { get; set; } //Importo da pagare
+        public double DebValue { get; set; }
         public string DebNote { get; set; }
     }
 }
