+++ Models/KnownMovements.cs
@@ -7,8 +7,8 @@ namespace PersonalFinance.Models
         public string Usr_OID { get; set; }
         public int ID { get; set; }
         public string KMType { get; set; }
-    public string KMTitle { get; set; }
-    public float KMValue { get; set; }
-    public string KMNote { get; set; }
+        public string KMTitle { get; set; }
+        public double KMValue { get; set; }
+        public string KMNote { get; set; }
     }
 }
